Implement "as though using ..." weapon type check
"Hollow Palm" introduces the "as though using a Quarterstaff" wording.
Current interpretation is that this is different from "count as using"
wording and only removes weapon type restrictions. This commit adds
support for that kind of behavior.

Note:
Because "WeaponData" mods are only processed in `Item.lua` and only for
actual items. This kind of mod will not work via `ModParser` when
modifying "Unarmed" weaponData