Fix display of some stats in gem tooltip
GGG added a new variant in the stat description files that is meant to be used only for gem quality tooltips.
We were not handling it properly before so if the same stat existed on the gem, the display for the stat was broken too.
e.g. Arc showed "Chains an additional time" instead of "Chains 9 times"
Now properly handled in the stat export and in stat describer