Add option to override condition values via `cfg`
It was kinda hard to tell `ModStore:GetCondition` to assume a condition
is `false` without changing the actual conditions on the actor, so I
added the opton to force the value via `cfg.overrideCond`.

I needed to force the condition because I wanted to tabulate mods that
do not apply, but are used as basis for the value of other mods. e.g.
Leech Elemental Damage *instead* of Physical Damage