Aura Magnitudes
I'm not 100% certain if this is the correct way to handle them or if we should just make them additive with curse / aura effect
`The Magnitudes of a Buff or Debuff are the values of the stats it applies to the target. A Buff or Debuff with higher magnitudes is more powerful.

Modifiers to the Magnitude of Buffs or Debuffs come from whoever applies it and are multiplicative with modifiers to the Effect the Buff or Debuff has on the target.`
This makes it sound to me that it's multiplicative with AuraEffect but all the stats internally use "curse_effect_+%" or "aura_effect_+%" so I don't really know anymore