Fix Damage conversion / gain calcs
calcConvertedDamage was using the wrong order of variables for the function so it wasn't taking into account gained damage if a skill had innate conversion on it.
E.g. Skill has 50% convert phys to fire and tree has 20% elemental as extra lightning, it was not finding the fire damage that had been converted so the ele as extra lightning was not giving any damage
Now that this was fixed though it was taking into account the proportion of damage converted from a skill when calculating gain as
e.g. 100% skill phys to fire stops mods like 20% phys as extra fire from working as there is no more base phys damage to scale from
Lastly was a bug that GGG fixed where damage mods that increased the maximum amount of a damage type (e.g. 30% more maximum phys) was affecting damage types after conversion
e.g. 100 to 100 phys damage skill with 30% more maximum phys and 100% skill phys to fire should have 100 to 100 fire damage, not 100 to130 that it was previously
I hope this is all the conversion issues fixed now