Fix Crash when hovering over some Tree nodes with Hulking Form allocated
Some mods use a table but don't have another mod nested inside the table like the code expects.
e.g. + Gem level mods
Some mods just a flag so instead of trying to identify if a mod is a flag, we only scale a line if the value returned is a number. That way `math.abs(newValue)` will always work

Also found that scaling Notables would break when we had a explode mod line as we had the value key named `chance`