Fix 'maxValue' not working for strength
`maxDiff` was initialized at `0`, which led to problems when the
difference in attribute ratio was nominally negative.

There is secondary issue that needs to be fixed related to this, but
at least this ensures that maximum value always applies as a limit.

The other issue is that target ratios should be recalculated if a
maximum value is reached because it changes the relative weights of the
remaining attributes. Will be done in separate commit