Import Quests from character ingame (#1558)
Added the logic to import quest rewards from your character in game
The quest rewards in the API are just a list of strings for the given rewards so we have to guess and match them to a quest
We loop through all the quests and see if the stat or options match 1 or more lines in the API list. If a match is made we remove those quest rewards from the API so that we don't enable multiple quests from 1 quest reward (e.g. +30 spirit is the reward from 2 different quests)
We also only rebuild the mod list if the config option is not already set from a previous import or manual change.

Co-authored-by: LocalIdentity <localidentity2@gmail.com>